+++ src/mplchart/primitives/__init__.py
@@ -11,3 +11,4 @@ from .lineplot import LinePlot
 from .areaplot import AreaPlot
 from .barplot import BarPlot
 from .zigzag import ZigZag
+from .pane import Pane