+++ src/mplchart/primitives/areaplot.py
@@ -47,8 +47,7 @@ class AreaPlot(Primitive):
 
     def plot_handler(self, prices, chart, ax=None):
         if ax is None:
-            target = self.target or chart.get_target(self.indicator)
-            ax = chart.get_axes(target)
+            ax = chart.get_axes(self.target)
 
         result = chart.calc_result(prices, self.indicator)